Improve display for big chainIDs

+++ src_features/signTx/ui_flow_signTx.c
@@ -222,9 +222,8 @@ void ux_approve_tx(bool fromPlugin) {
         ux_approval_tx_flow[step++] = &ux_approval_nonce_step;
     }
 
-    uint32_t chain_id = get_chain_id();
+    uint64_t chain_id = get_chain_id();
     if (chainConfig->chainId == ETHEREUM_MAINNET_CHAINID && chain_id != chainConfig->chainId) {
-        // TODO: do we need the `&&` above?
         ux_approval_tx_flow[step++] = &ux_approval_network_step;
     }
